requirements.txt
Flask==2.3.2
pymongo==4.4.0
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
python-dotenv==1.0.0
dnspython==2.4.2
gunicorn==20.1.0
